Projects/querygen/requirements.txt
fastapi==0.116.1
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
setuptools==80.9.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
streamlit==1.49.1
uvicorn==0.35.0
